fix(trusty-review): parse {"indexes":[...]} envelope in list_indexes so auto-derive resolves the real index (closes #672)
Two bugs prevented resolve_index() from ever finding the correct index:

1. list_indexes() deserialised the daemon response as Vec<IndexInfo> directly,
   but trusty-search returns {"indexes":[...]} — causing `invalid type: map,
   expected a sequence`. resolve_index() caught this as a daemon error and fell
   back to "main" every time.

2. The request targeted GET /indexes (bare), but root_path is only present in
   the GET /indexes?details=true response. Without root_path, best_matching_index
   finds no match even when the index is registered.

Fix:
- Add ListIndexesResponse wrapper struct (#[derive(Deserialize)]) and unwrap .indexes.
- Change URL to /indexes?details=true so root_path is populated.
- Split inline tests to search_client_tests.rs to keep search_client.rs under
  the 500-line cap (#610).
- Add regression tests: list_indexes_parses_daemon_envelope (real envelope payload)
  and list_indexes_bare_array_is_rejected (documents the new shape contract).

References: #661, #670.
